issue #4660: Warning fix in CCLabel.cpp and ConvertUTF.c

+++ cocos/2d/CCLabel.cpp
@@ -1193,7 +1193,7 @@ void Label::computeStringNumLines()
 
 int Label::getStringLength() const
 {
-    return _currentUTF16String.length();
+    return static_cast<int>(_currentUTF16String.length());
 }
 
 // RGBA protocol
